requirements.txt
requests==2.13.0
Pillow==5.4.1
